Extract safe_path into shared/path_utils.py
Single source of truth for the path traversal check. Both
sonar-update-center-release/update.py and test-fixtures/jira/jira_client.py
now import from it via sys.path.
